Splits hook helpers into several modules so to avoid curcular imports
