+++ packages/react/src/helpers/__tests__/operations.spec.ts
@@ -1,50 +1,9 @@
-import { ApolloError, QueryResult as ApolloQueryResult } from '@apollo/client';
-import { UnspecifiedError } from '@lens-protocol/api-bindings';
 import { Deferred, failure, Result, success } from '@lens-protocol/shared-kernel';
 import { act, renderHook } from '@testing-library/react';
 
-import {
-  OperationHandler,
-  PaginatedQueryData,
-  QueryData,
-  useOperation,
-  usePaginatedReadResult,
-  useReadResult,
-} from '../helpers';
+import { OperationHandler, useOperation } from '../operations';
 
-describe(`Given the hook helpers`, () => {
-  describe(`when creating an hook with the ${useReadResult.name} helper`, () => {
-    it(`should wrap any error into ${UnspecifiedError.name}`, async () => {
-      const cause = new ApolloError({ graphQLErrors: [] });
-      const queryResult = { error: cause, data: undefined, loading: false } as ApolloQueryResult<
-        QueryData<void>
-      >;
-      const { result } = renderHook(() => useReadResult(queryResult));
-
-      expect(result.current).toMatchObject({
-        error: expect.any(UnspecifiedError),
-        data: undefined,
-        loading: false,
-      });
-    });
-  });
-
-  describe(`when creating an hook with the ${usePaginatedReadResult.name} helper`, () => {
-    it(`should wrap any error into ${UnspecifiedError.name}`, async () => {
-      const cause = new ApolloError({ graphQLErrors: [] });
-      const queryResult = { error: cause, data: undefined, loading: false } as ApolloQueryResult<
-        PaginatedQueryData<void>
-      >;
-      const { result } = renderHook(() => usePaginatedReadResult(queryResult));
-
-      expect(result.current).toMatchObject({
-        error: expect.any(UnspecifiedError),
-        data: undefined,
-        loading: false,
-      });
-    });
-  });
-
+describe(`Given the operation hook helpers`, () => {
   describe(`when creating an hook with the ${useOperation.name} helper`, () => {
     it('should call the provided handler function with the arguments passed to the execute method', async () => {
       const fn: OperationHandler<void, never, unknown[]> = jest.fn().mockResolvedValue(success());
