+++ packages/react/src/helpers/operations.ts
@@ -1,0 +1,46 @@
+import { IEquatableError, PromiseResult } from '@lens-protocol/shared-kernel';
+import { useState } from 'react';
+
+export type OperationHandler<
+  TResult,
+  TError extends IEquatableError,
+  TArgs extends unknown[] = never,
+> = (...args: TArgs) => PromiseResult<TResult, TError>;
+
+export type Operation<
+  TResult,
+  TError extends IEquatableError = never,
+  TArgs extends unknown[] = [],
+> = {
+  error: TError | undefined;
+  execute: (...args: TArgs) => PromiseResult<TResult, TError>;
+  isPending: boolean;
+};
+
+export function useOperation<TResult, TError extends IEquatableError, TArgs extends unknown[]>(
+  handler: OperationHandler<TResult, TError, TArgs>,
+): Operation<TResult, TError, TArgs> {
+  const [isPending, setIsPending] = useState(false);
+  const [error, setError] = useState<TError | undefined>(undefined);
+
+  return {
+    error,
+    execute: async (...args: TArgs) => {
+      setError(undefined);
+      setIsPending(true);
+
+      try {
+        const result = await handler(...args);
+
+        if (result.isFailure()) {
+          setError(result.error);
+        }
+
+        return result;
+      } finally {
+        setIsPending(false);
+      }
+    },
+    isPending,
+  };
+}
