+++ packages/react/src/index.tsx
@@ -11,7 +11,7 @@ export * from './wallet';
 export * from './notifications';
 export * from './experimental';
 
-export type { ReadResult, PaginatedReadResult } from './helpers';
+export type { ReadResult, PaginatedReadResult } from './helpers/reads';
 export { isValidHandle } from '@lens-protocol/api-bindings';
 export type {
   StorageSubscription,
