+++ packages/react/src/modules/useEnabledModules.ts
@@ -1,6 +1,6 @@
 import { useEnabledModulesQuery } from '@lens-protocol/api-bindings';
 
-import { useReadResult } from '../helpers';
+import { useReadResult } from '../helpers/reads';
 import { useSharedDependencies } from '../shared';
 
 export function useEnabledModules() {
