+++ packages/react/src/notifications/useNotifications.ts
@@ -9,7 +9,7 @@ import {
   useNotificationsQuery,
 } from '@lens-protocol/api-bindings';
 
-import { PaginatedArgs, PaginatedReadResult, usePaginatedReadResult } from '../helpers';
+import { PaginatedArgs, PaginatedReadResult, usePaginatedReadResult } from '../helpers/reads';
 import { useSharedDependencies } from '../shared';
 
 type UseNotificationsArgs = PaginatedArgs<{
