+++ packages/react/src/profile/useActiveProfileIdentifier.ts
@@ -1,6 +1,6 @@
 import { ProfileIdentifier } from '@lens-protocol/domain/use-cases/profile';
 
-import { ReadResult } from '../helpers';
+import { ReadResult } from '../helpers/reads';
 import { ApplicationsState, useAppState } from '../lifecycle/adapters/ApplicationPresenter';
 import { useActiveProfileIdentifierVar } from './adapters/ActiveProfilePresenter';
