+++ packages/react/src/profile/useActiveProfileSwitch.ts
@@ -1,7 +1,7 @@
 import { ProfileId } from '@lens-protocol/domain/entities';
 import { success } from '@lens-protocol/shared-kernel';
 
-import { Operation, useOperation } from '../helpers';
+import { Operation, useOperation } from '../helpers/operations';
 import { useSwitchActiveProfileController } from './adapters/useSwitchActiveProfileController';
 
 export type ActiveProfileSwitchOperation = Operation<void, never, [ProfileId]>;
