+++ packages/react/src/profile/useExploreProfiles.ts
@@ -1,6 +1,6 @@
 import { ProfileFragment, useExploreProfilesQuery } from '@lens-protocol/api-bindings';
 
-import { PaginatedReadResult, PaginatedArgs, usePaginatedReadResult } from '../helpers';
+import { PaginatedReadResult, PaginatedArgs, usePaginatedReadResult } from '../helpers/reads';
 import { useSharedDependencies } from '../shared';
 
 type UseFeedArgs = PaginatedArgs<{
