+++ packages/react/src/profile/useProfileFollowing.ts
@@ -1,6 +1,6 @@
 import { FollowingFragment, useProfileFollowingQuery } from '@lens-protocol/api-bindings';
 
-import { PaginatedArgs, PaginatedReadResult, usePaginatedReadResult } from '../helpers';
+import { PaginatedArgs, PaginatedReadResult, usePaginatedReadResult } from '../helpers/reads';
 import { useSharedDependencies } from '../shared';
 
 type UseProfileFollowingArgs = PaginatedArgs<{
