+++ packages/react/src/profile/useProfilesToFollow.ts
@@ -1,6 +1,6 @@
 import { useProfilesToFollowQuery } from '@lens-protocol/api-bindings';
 
-import { useReadResult } from '../helpers';
+import { useReadResult } from '../helpers/reads';
 import { useSharedDependencies } from '../shared';
 
 export function useProfilesToFollow() {
