+++ packages/react/src/publication/useComments.ts
@@ -1,6 +1,6 @@
 import { useCommentsQuery, CommentWithFirstCommentFragment } from '@lens-protocol/api-bindings';
 
-import { PaginatedArgs, PaginatedReadResult, usePaginatedReadResult } from '../helpers';
+import { PaginatedArgs, PaginatedReadResult, usePaginatedReadResult } from '../helpers/reads';
 import { useSharedDependencies } from '../shared';
 import { createPublicationMetadataFilters, PublicationMetadataFilters } from './filters';
