+++ packages/react/src/publication/useProfilePublicationsForSale.ts
@@ -3,7 +3,7 @@ import {
   useProfilePublicationsForSaleQuery,
 } from '@lens-protocol/api-bindings';
 
-import { PaginatedArgs, PaginatedReadResult, usePaginatedReadResult } from '../helpers';
+import { PaginatedArgs, PaginatedReadResult, usePaginatedReadResult } from '../helpers/reads';
 import { useSharedDependencies } from '../shared';
 import { DEFAULT_PAGINATED_QUERY_LIMIT } from '../utils';
