+++ packages/react/src/publication/useWhoCollectedPublication.ts
@@ -1,6 +1,6 @@
 import { useWhoCollectedPublicationQuery, WalletFragment } from '@lens-protocol/api-bindings';
 
-import { PaginatedReadResult, PaginatedArgs, usePaginatedReadResult } from '../helpers';
+import { PaginatedReadResult, PaginatedArgs, usePaginatedReadResult } from '../helpers/reads';
 import { useSharedDependencies } from '../shared';
 import { DEFAULT_PAGINATED_QUERY_LIMIT } from '../utils';
