+++ packages/react/src/revenue/useProfilePublicationRevenue.ts
@@ -4,7 +4,7 @@ import {
   useProfilePublicationRevenueQuery,
 } from '@lens-protocol/api-bindings';
 
-import { PaginatedArgs, PaginatedReadResult, usePaginatedReadResult } from '../helpers';
+import { PaginatedArgs, PaginatedReadResult, usePaginatedReadResult } from '../helpers/reads';
 import { useSharedDependencies } from '../shared';
 import { DEFAULT_PAGINATED_QUERY_LIMIT } from '../utils';
