+++ packages/react/src/transactions/useCollect.ts
@@ -9,7 +9,7 @@ import {
   WalletConnectionError,
 } from '@lens-protocol/domain/entities';
 
-import { Operation, useOperation } from '../helpers';
+import { Operation, useOperation } from '../helpers/operations';
 import { useCollectController } from './adapters/useCollectController';
 
 export type UseCollectArgs = {
