+++ packages/react/src/transactions/useCreateMirror.ts
@@ -10,7 +10,7 @@ import {
   WalletConnectionError,
 } from '@lens-protocol/domain/entities';
 
-import { Operation, useOperation } from '../helpers';
+import { Operation, useOperation } from '../helpers/operations';
 import { useCreateMirrorController } from './adapters/useCreateMirrorController';
 
 export type UseCreateMirrorArgs = {
