+++ packages/react/src/transactions/useUpdateProfileDetails.ts
@@ -8,7 +8,7 @@ import {
 import { UpdateProfileDetailsRequest } from '@lens-protocol/domain/use-cases/profile';
 import { failure, PromiseResult } from '@lens-protocol/shared-kernel';
 
-import { Operation, useOperation } from '../helpers';
+import { Operation, useOperation } from '../helpers/operations';
 import { FailedUploadError } from './adapters/IMetadataUploader';
 import { MetadataUploadHandler } from './adapters/MetadataUploadHandler';
 import { useUpdateProfileDetailsController } from './adapters/useUpdateProfileDetailsController';
